get rid of some cyclic import
